@@ -13,16 +13,18 @@ license = {text = "MIT License"}
 authors = [
   {name = "Scott Lundberg", email = "slund1@cs.washington.edu"},
 ]
-requires-python = ">=3.9"
+requires-python = ">=3.11"
 dependencies = [
-  'numpy',
+  'numpy>=2',
   'scipy',
   'scikit-learn',
   'pandas',
   'tqdm>=4.27.0',
   'packaging>20.9',
   'slicer==0.0.8',
-  'numba>=0.54',
+  'numba==0.63.0b1; python_version >= "3.14"',  # todo: remove once numba officially releases support for 3.14
+  'llvmlite==0.46.0b1; python_version >= "3.14"',  # todo: remove once llvmlite officially releases support for 3.14
+  'numba>=0.54; python_version < "3.14"',
   'cloudpickle',
   'typing-extensions',
 ]
@@ -31,11 +33,10 @@ classifiers = [
   "Operating System :: POSIX",
   "Operating System :: Unix",
   "Operating System :: MacOS",
-  "Programming Language :: Python :: 3.9",
-  "Programming Language :: Python :: 3.10",
   "Programming Language :: Python :: 3.11",
   "Programming Language :: Python :: 3.12",
   "Programming Language :: Python :: 3.13",
+  "Programming Language :: Python :: 3.14",
   "Intended Audience :: Information Technology",
   "Intended Audience :: Science/Research",
   "Topic :: Scientific/Engineering",
@@ -77,22 +78,22 @@ test = [
   "pytest-cov",
   "xgboost",
   "lightgbm",
-  "catboost; python_version < '3.13'",  # TODO: pending 3.13 support
-  "gpboost<=1.6.1",  # TODO: can be relaxed once https://github.com/fabsig/GPBoost/issues/165 is solved
+  'catboost; python_version < "3.14"',  # TODO: pending 3.14 support
+  'gpboost; python_version < "3.14"',  # TODO: pending 3.14 support
   "ngboost",
   "pyspark",
   "pyod",
-  "transformers < 4.54.0; python_version < '3.13'",  # TODO: pending 3.13 support, fix once tests pass for 4.54.0
-  "tf-keras; python_version < '3.13'",
-  "protobuf==3.20.3",  # See GH #3046
-  "torch; python_version < '3.13'",  # TODO: pending 3.13 support
-  "torchvision; python_version < '3.13'",  # TODO: pending 3.13 support
-  "tensorflow; python_version < '3.13'",  # TODO: pending 3.13 support
+  "transformers < 4.54.0",  # TODO: fix once tests pass for 4.54.0
+  "tf-keras",
+  "protobuf",  # See GH #3046
+  'torch; python_version < "3.14"',  # TODO: pending 3.14 support
+  'torchvision; python_version < "3.14"',  # TODO: pending 3.14 support
+  'tensorflow; python_version < "3.14"',  # TODO: pending 3.14 support
   "sentencepiece",
   "opencv-python",
   # Constraint to prevent the combination of tf<2.15 and numpy>=2.0.
   # See GH #3707, #3768, 3922
-  "numpy<2.0",
+  "numpy>=2.0",
   "scikit-learn<=1.6.1",  # needed since causalml doesn't support scikit-learn 1.7.0 and above as of 2025-06-11
   "causalml",
   "selenium",  # needed to test the javascript based plots
