@@ -7,7 +7,7 @@ from __future__ import annotations
 
 import dataclasses
 from contextlib import contextmanager
-from typing import TypedDict, Union
+from typing import TypeAlias, TypedDict
 
 import numpy as np
 from typing_extensions import Unpack
@@ -16,17 +16,19 @@ from ..utils._exceptions import InvalidStyleOptionError
 from . import colors
 
 # Type hints, adapted from matplotlib.typing
-RGBColorType = Union[tuple[float, float, float], str]
-RGBAColorType = Union[
-    str,  # "none" or "#RRGGBBAA"/"#RGBA" hex strings
-    tuple[float, float, float, float],
+RGBColorType = tuple[float, float, float] | str
+RGBAColorType: TypeAlias = (
+    str  # "none" or "#RRGGBBAA"/"#RGBA" hex strings
+    | tuple[float, float, float, float]
+    |
     # 2 tuple (color, alpha) representations, not infinitely recursive
     # RGBColorType includes the (str, float) tuple, even for RGBA strings
-    tuple[RGBColorType, float],
+    tuple[RGBColorType, float]
+    |
     # (4-tuple, float) is odd, but accepted as the outer float overriding A of 4-tuple
-    tuple[tuple[float, float, float, float], float],
-]
-ColorType = Union[RGBColorType, RGBAColorType, np.ndarray]
+    tuple[tuple[float, float, float, float], float]
+)
+ColorType: TypeAlias = RGBColorType | RGBAColorType | np.ndarray
 
 
 # TODO: Use dataclass(kw_only=True) when we drop Python 3.9
